+++ src/intel/compiler/ibc_live_intervals.h
@@ -104,7 +104,7 @@ typedef struct {
 void
 ibc_live_intervals_reg_ref_chunks(const ibc_live_intervals *live,
                                   const ibc_reg_ref *ref,
-                                  int8_t num_bytes, int8_t num_comps,
+                                  int num_bytes, int num_comps,
                                   uint8_t simd_group, uint8_t simd_width,
                                   BITSET_WORD *chunks);
